time/Clock: BUG-FIX

+++ src/chronotext/time/Clock.cpp
@@ -1,6 +1,6 @@
 /*
  * THE NEW CHRONOTEXT TOOLKIT: https://github.com/arielm/new-chronotext-toolkit
- * COPYRIGHT (C) 2012, ARIEL MALKA ALL RIGHTS RESERVED.
+ * COPYRIGHT (C) 2012-2014, ARIEL MALKA ALL RIGHTS RESERVED.
  *
  * THE FOLLOWING SOURCE-CODE IS DISTRIBUTED UNDER THE MODIFIED BSD LICENSE:
  * https://github.com/arielm/new-chronotext-toolkit/blob/master/LICENSE.md
@@ -82,7 +82,7 @@ namespace chronotext
         return mst + ((state == STOPPED) ? 0 : (timeBase->getTime() - tbst) * rate);
     }
     
-    void Clock::setTime(int now)
+    void Clock::setTime(double now)
     {
         if (state == STOPPED)
         {
