+++ src/chronotext/time/Clock.h
@@ -1,6 +1,6 @@
 /*
  * THE NEW CHRONOTEXT TOOLKIT: https://github.com/arielm/new-chronotext-toolkit
- * COPYRIGHT (C) 2012, ARIEL MALKA ALL RIGHTS RESERVED.
+ * COPYRIGHT (C) 2012-2014, ARIEL MALKA ALL RIGHTS RESERVED.
  *
  * THE FOLLOWING SOURCE-CODE IS DISTRIBUTED UNDER THE MODIFIED BSD LICENSE:
  * https://github.com/arielm/new-chronotext-toolkit/blob/master/LICENSE.md
@@ -48,7 +48,7 @@ namespace chronotext
         virtual void start();
         virtual void stop();
         virtual double getTime();
-        virtual void setTime(int now);
+        virtual void setTime(double now);
         virtual int getState();
         virtual void setRate(double factor);
         virtual void restart();
